# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is gopa-server. Use these APIs to manage different groups, roles of GOPA",
        "title": "Gopa-server Swagger Documentation",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/application/delete": {
            "post": {
                "security": [
                    {
                        "Token": []
                    }
                ],
                "description": "Delete this application from OPA. This method will delete ALL RELATED INFORMATION from database.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "application"
                ],
                "summary": "ApplicationDelete",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Project",
                        "name": "project",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Role",
                        "name": "role",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "form",
                        "name": "form",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.ApplicationDeleteForm"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    }
                }
            }
        },
...
        "/api/v1/project/delete": {
            "post": {
                "security": [
                    {
                        "Token": []
                    }
                ],
                "description": "Delete a project",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "project"
                ],
                "summary": "ProjectDelete",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Project",
                        "name": "project",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Role",
                        "name": "role",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "form",
                        "name": "form",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.ProjectDeleteForm"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    }
                }
            }
        },
...
        "/api/v1/projectResource/delete": {
            "post": {
                "security": [
                    {
                        "Token": []
                    }
                ],
                "description": "Delete a projectRole from a resource.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "projectResource"
                ],
                "summary": "ProjectResourceDelete",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Project",
                        "name": "project",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Role",
                        "name": "role",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "form",
                        "name": "form",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.ApplicationDeleteForm"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    }
                }
            }
        },
...
        "/api/v1/projectRole/delete": {
            "post": {
                "security": [
                    {
                        "Token": []
                    }
                ],
                "description": "Delete a role from a given project",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "projectRole"
                ],
                "summary": "ProjectRoleDelete",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Project",
                        "name": "project",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Role",
                        "name": "role",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "form",
                        "name": "form",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.RoleDeleteFromGroupForm"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    }
                }
            }
        },
...
        "/api/v1/rego/delete": {
            "post": {
                "security": [
                    {
                        "Token": []
                    }
                ],
                "description": "Delete a rego file by given path",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rego"
                ],
                "summary": "RegoDelete",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Project",
                        "name": "project",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Role",
                        "name": "role",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "form",
                        "name": "form",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.DeletedRegoDocumentForm"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    }
                }
            }
        },
...
        "/api/v1/role/delete": {
            "post": {
                "security": [
                    {
                        "Token": []
                    }
                ],
                "description": "Delete a role",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "role"
                ],
                "summary": "RoleDelete",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Project",
                        "name": "project",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Role",
                        "name": "role",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "form",
                        "name": "form",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.DeletedRoleForm"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    }
                }
            }
        },
...
        "/api/v1/user/delete": {
            "post": {
                "security": [
                    {
                        "Token": []
                    }
                ],
                "description": "Delete a user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "UserDelete",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Project",
                        "name": "project",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Role",
                        "name": "role",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "form",
                        "name": "form",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.UserDeleteForm"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "api.ApplicationDeleteForm": {
            "type": "object",
            "properties": {
                "application_name": {
                    "type": "string"
                }
            }
        },
        "api.DeletedRegoDocumentForm": {
            "type": "object",
            "properties": {
                "path": {
                    "type": "string"
                }
            }
        },
        "api.DeletedRoleForm": {
            "type": "object",
            "properties": {
                "role_name": {
                    "type": "string"
                }
            }
        },
...
        "api.ProjectDeleteForm": {
            "type": "object",
            "properties": {
                "project_name": {
                    "type": "string"
                }
            }
        },
...
        "api.RoleDeleteFromGroupForm": {
            "type": "object",
            "properties": {
                "project_name": {
                    "type": "string"
                },
                "role_name": {
                    "type": "string"
                }
            }
        },
...
        "api.UserDeleteForm": {
            "type": "object",
            "properties": {
                "username": {
                    "type": "string"
                }
            }
        },
        "handler.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "Token": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}